+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.3.0] - 2019-05-09
+
+### Added
+* **rdg** action
+    * Find RDCMan.settings and linked .RDG files, or take a given /target .RDG/RDCMan.settings file/folder, and decrypt passwords given a /pvk, GUID key lookup table, or CryptUnprotectData (with /unprotect).
+* **blob** action
+    * Describe a supplied DPAPI binary blob, optionally decryption the blob with masterkey GUID lookups or a PVK masterkey decryption
+
+### Changed
+* Added IsTextUnicode() for vault/credential/blob decryption display, showing hex if unicode is detected
+* Added /target:C:\FOLDER\ option for the **masterkeys** function, for offline masterkey decryption
+* Updated README
+
+
 ## [1.2.0] - 2019-03-24 (Troopers edition ;)
 
 ### Added